MBS-14209: Add lesarchivesduspectacle.net to the otherdbs whitelist
A database of mostly French stage productions, including
opera and theatre.

We support three types of links:
/p for people (composers, performers, engineers)
/oe for works (such as operas and plays)
/o for organizations (like venues, but also festivals)

A fourth type of link, /s (for spectacles), does not seem to
match any of our levels very well, since it's not quite
a tour series but not quite an event.
I'm blocking these for now.