Gate "local accounts" behind DBDefs setting
With accounts moving to MetaBrainz and authentication happening via OAuth,
we'll preserve the ability to use local accounts in development in order to not
require a MetaBrainz development server.

This is gated behind the `LOCAL_ACCOUNTS_ENABLED` DBDefs setting. It's enabled
in this commit, but will be disabled by default once MetaBrainz authentication
is implemented.