MBS-13257: Add Hörspielforscher to other databases whitelist
[Die Hörspielforscher] is a German audio drama database which covers
audio drama releases (and related music releases), the involved artists
and labels, as well as detailed cuesheets of the featured music.

[Die Hörspielforscher]: https://hoerspielforscher.de/

All relevant URLs follow the pattern
`https://hoerspielforscher.de/kartei/ENTITY_TYPE?detail=ID`.
URL fragments and most query parameters can and should be discarded to
obtain the most *canonical* URLs. Except for the ID from the `detail`
parameter, they only encode positioning and state of the page.

When these are omitted, the default tab and filters are used as it is
desired most of the time. Only `person` pages could be considered an
exception if e.g. a person is mainly a musician ([music tab]) but also
has a few voice actor credits which are then [shown by default].
But I guess that's ok for a database which is mainly about audio dramas.

[music tab]: https://hoerspielforscher.de/kartei/person?detail=668&detailSub=musiker
[shown by default]: https://hoerspielforscher.de/kartei/person?detail=668

Most of the type mappings are straightforward, the only oddity is that
a `person` can be a human/artist as well as a company/label.
All persons have `person` pages while bands and musicians (additionally)
have `interpret` pages.
A release can have an `album` (music) or a `hoerspiel` (audio drama)
page, both share the same IDs and could technically be standardized.
Imprints have `label` pages, release series have `serie` pages and music
recordings have `musik` pages.

I have omitted a few other types which don't fit into the MB schema:

* `suite`: series/collection of recordings which were recorded by the
  same artists during a limited time span, which ranges from a session
  to multiple years
* `audio`: cuesheet of the featured music, can be shared by multiple
  releases, but not equivalent to a release group (as both edited
  releases and mediums with sides will have a different cuesheet)
* `katalog`: catalog number range, can be shared by related imprints