MBS-14213: Drop handing of pomus.net links
This went away in 2022. It really does not look like it is coming
back so we can just drop the special handling and drop it
from sidebars (the relationships should be bot-marked as ended).

I'm not removing the entry from the otherdatabases match
in UrlCleanup since IIUC that would make the ended links
show as errors, and it would probably be confusing.